WIP - Created Test: ParseTakesInBytesOfARequestWithNoHeadersAndReturnsARequest
