+++ src/ProtocolHandler/DataParser/BasicDataParser.cs
@@ -1,0 +1,10 @@
+namespace Chorizo.ProtocolHandler.DataParser
+{
+    public class BasicDataParser:IDataParser
+    {
+        public Request Parse(byte[] startLineAndHeadersBytes)
+        {
+            throw new System.NotImplementedException();
+        }
+    }
+}
