+++ src/ProtocolHandler/DataParser/IDataParser.cs
@@ -1,0 +1,7 @@
+namespace Chorizo.ProtocolHandler.DataParser
+{
+    public interface IDataParser
+    {
+        Request Parse(byte[] startLineAndHeadersBytes);
+    }
+}
